fix: map tinyint columns AdminLevel and Leave to byte in User model
Agent-Logs-Url: https://github.com/rmarsy/sh/sessions/4ac1ee06-5fa0-40d7-b17a-1392524b2dcb

+++ Sh.Api/Models/User.cs
@@ -25,13 +25,13 @@ public class User
     public bool Admin { get; set; }
 
     [Column("adminlevel")]
-    public int AdminLevel { get; set; }
+    public byte AdminLevel { get; set; }
 
     [Column("status")]
     public short Status { get; set; }
 
     [Column("leave")]
-    public int Leave { get; set; }
+    public byte Leave { get; set; }
 
     [Column("leavedate")]
     public DateTime? LeaveDate { get; set; }
